06-form-demo/agent/pyproject.toml
[project]
name = "lemonslice-form-demo-agent"
version = "0.1.0"
description = "LiveKit Agents worker for the LemonSlice form demo (voice SDR + sidebar)"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[lemonslice]>=1.2.0,<2",
    "livekit-plugins-silero>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[tool.uv]
package = false
